# apps/delivery-service/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/drivers": {
      "get": {
        "operationId": "DriversController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all drivers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all drivers (admin only)",
        "tags": [
          "Drivers"
        ]
      },
      "post": {
        "operationId": "DriversController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Driver created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new driver",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/drivers/available": {
      "get": {
        "operationId": "DriversController_findAvailable",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of available drivers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get available drivers",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/drivers/me": {
      "get": {
        "operationId": "DriversController_getMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Driver profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current driver profile",
        "tags": [
          "Drivers"
        ]
      },
      "put": {
        "operationId": "DriversController_updateMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current driver profile",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/drivers/{id}": {
      "get": {
        "operationId": "DriversController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get driver by ID (admin only)",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/drivers/me/location": {
      "patch": {
        "operationId": "DriversController_updateLocation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update driver location",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/drivers/me/availability": {
      "patch": {
        "operationId": "DriversController_setAvailability",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set driver availability status",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/drivers/me/deactivate": {
      "patch": {
        "operationId": "DriversController_deactivateMyAccount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Driver deactivated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate current driver account",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/drivers/{id}/deactivate": {
      "patch": {
        "operationId": "DriversController_deactivateDriver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver deactivated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a driver (admin only)",
        "tags": [
          "Drivers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Delivery Service API",
    "description": "A modular monolith delivery service system with Clean Architecture, CQRS, and Event-Driven patterns",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DriverLocationDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "example": 40.7128
          },
          "longitude": {
            "type": "number",
            "example": -74.006
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "latitude",
          "longitude",
          "updatedAt"
        ]
      },
      "DriverResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "userId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "vehicleType": {
            "type": "string",
            "example": "Sedan"
          },
          "licensePlate": {
            "type": "string",
            "example": "ABC-1234"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "busy",
              "offline"
            ],
            "example": "available"
          },
          "currentLocation": {
            "$ref": "#/components/schemas/DriverLocationDto"
          },
          "rating": {
            "type": "number",
            "example": 4.8
          },
          "totalDeliveries": {
            "type": "number",
            "example": 150
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "vehicleType",
          "licensePlate",
          "status",
          "rating",
          "totalDeliveries",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateDriverDto": {
        "type": "object",
        "properties": {
          "vehicleType": {
            "type": "string",
            "minLength": 2
          },
          "licensePlate": {
            "type": "string",
            "minLength": 2
          }
        },
        "required": [
          "vehicleType",
          "licensePlate"
        ]
      },
      "UpdateDriverDto": {
        "type": "object",
        "properties": {
          "vehicleType": {
            "type": "string",
            "minLength": 2
          },
          "licensePlate": {
            "type": "string",
            "minLength": 2
          }
        },
        "required": [
          "vehicleType",
          "licensePlate"
        ]
      },
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "exclusiveMinimum": false,
            "maximum": 90,
            "exclusiveMaximum": false
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "exclusiveMinimum": false,
            "maximum": 180,
            "exclusiveMaximum": false
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "SetAvailabilityDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "available",
              "busy",
              "offline"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}